requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
future==0.18.2
gspread==3.1.0
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
line-bot-sdk==1.16.0
oauth2client==4.1.3
psycopg2==2.8.4
APScheduler==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.7
mysql-connector==2.2.9        
mysql-connector-python==8.0.19 
PyMySQL==0.9.3
